Modified to ensure topological entites are spatially continuous
Now uses a modified flood fill algorithm to tag spatially continous
volumes. Appends a character to each topology code. If two separated
volumes have the same topology code then different letters are appended
to to each the topology code of each volume.

Note that this algorithm requires an enlarged stack space. On windows
this requires compiling with -Wl,--stack,100000000. The application
automatically resizes the stack to 1Gb on linux systems.