Migrated Experiment classes to submodule
Pulled Experiment classes into separate files and put them in a
submodule called pynoddy.experiment. Note that the root pynoddy
directory now needs to be on the pythonpath for things to work (though I
think it did before also)!

I have tested the experiment class & my ResolutionTest class, but not
the SensitivityAnalysis class as I don't know what it is supposed to do.