Changed topology.c to run single files only
Changed topology.c to only run a single model at a time, so that python
can call it straight after running noddy. This improves the flow of
multi-threaded workflows, and allows topologies to be calculated for
files that don't end in _000x (which was annoying).

topology.c was also simplified by using a loop to calculate neighbour
relations. It now also considers diagonal voxels as neighbours.

The MonteCarlo class has been simplified so that the topology code is
automatically run if sim_type == "TOPOLOGY"