Change structure of event params dictionary
Changed event params dict so that event tuples can now be passed. Event
tuples are generated from a params csv file if event id's are separated
by the '|' character.

Old versions of the params files *should* still work.

The purpose of this change is so different parameters can be randomly
sampled between models, but retain the same value within models. This is
necessary if, for example, a fault is represented twice, once during
fault formation and once during inversion.